Updated expected type of `ML::pred` bank entries.

+++ app/src/main/groovy/org/jlab/analysis/Kinematics.groovy
@@ -90,7 +90,7 @@ public class Kinematics {
 		Schema schema = reader.getSchemaFactory().getSchema("ML::pred");
         Bank bank     = new Bank(schema);
         event.read(bank);
-		pred = (double) bank.getFloat("pred",0);
+		pred = (double) bank.getDouble("pred",0);
 		return pred; };
 
     protected ConfigVar _getMLLabel = (ConfigVar)(HipoReader reader, Event event) -> {
@@ -98,7 +98,7 @@ public class Kinematics {
 		Schema schema = reader.getSchemaFactory().getSchema("ML::pred");
         Bank bank     = new Bank(schema);
         event.read(bank);
-		label = (double) bank.getFloat("label",0);
+		label = (double) bank.getDouble("label",0);
 		return label; };
 
     protected ConfigVar _getHelicityMC = (ConfigVar)(HipoReader reader, Event event) -> {
